Make card selectable on main page.
